% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options_detach.R
\name{options_detach}
\alias{options_detach}
\title{Detachment based on options}
\usage{
options_detach(cleanup = FALSE, target)
}
\arguments{
\item{cleanup}{Logical value, default \code{FALSE}. This can be
used to get rid of all the stuff that \code{options_attach}
added to the search-path.  The \code{target}-argument will be
ignored when \code{cleanup} is given as \code{TRUE}, and does
not need to be specified in that case.}

\item{target}{A character string that has to be one of the options
set by \code{options_attach}.}
}
\value{
The result of this function will be that components added
    to the search-path by \code{options_attach} will be detached.
}
\description{
This function does the opposite of \code{options_attach}, that is to
say, it can use the options created by \code{options_attach} to
detach the specified components from the search-path.
}

