% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_cleanup.R
\name{uc_cleanup}
\alias{uc_cleanup}
\title{Clean up workspace and the search-path after using \code{under_construction}}
\usage{
uc_cleanup()
}
\description{
The function \code{under_construction} adds stuff to the
search-path, and it's also required with a cleanup afterwards.  This
function checks if it has been created a global option with the
name \code{uc_temp_attached_env}.  If that's the case, then it will
proceed to detach the environments mentioned there, and moreover
also remove the corresponding objects from the workspace if they
are still present there.
}

