% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_extractor.R
\name{uc_extractor}
\alias{uc_extractor}
\title{Extract information from a package.}
\usage{
uc_extractor(.package_name, .attach = FALSE)
}
\arguments{
\item{.package_name}{The name of the package to investigate.}

\item{.attach}{Logical value, default \code{FALSE}, that decides if
the extracted values should be added to the search-path.}
}
\value{
Depending on \code{.attach}.  When \code{FALSE}, a list
    with two environments, one containing any non-exported objects
    from the package and one containing copies of stuff that was
    imported into the package (stuff already available in the
    search-path will be ignored).  When \code{TRUE} these
    environments are added to the end of the search-path., and
    nothing is returned to the work-flow.  Note that
    \code{uc_cleanup} can be used to get rid of stuff again.
}
\description{
In order to simplify the investigation by
\code{under_construction}, it might in many cases be preferable to
extract from a package the non-exported objects, and in addition
get our hands the functions that has been imported into the package.
}

