% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_help.R
\name{uc_help}
\alias{uc_help}
\title{Help function for \code{\link{under_construction}}}
\usage{
uc_help(.uc_fun, high_details_level = FALSE, clean_workspace = FALSE,
  script_files = "")
}
\arguments{
\item{.uc_fun}{The function we want to investigate.}

\item{high_details_level}{Logic value, default \code{FALSE}, to be
delivered to \code{under_construction}.}

\item{clean_workspace}{Logic value, default \code{FALSE}, to be
delivered to \code{under_construction}.}

\item{script_files}{Character vector, default \code{""} (i.e. an
empty string), to be delivered to \code{under_construction}.}
}
\value{
This function returns some lines of text, that gives three
    chunks of code to simplify the setup when working with the
    function \code{under_construction}, i.e. it will create an
    argument list containing all the arguments of \code{.fun} with
    the default values included, and this list will then be used to
    feed arguments to the \code{dotsMethods} argument of
    \code{under_construction}.  Code needed for an inspection of
    the result is also included.
}
\description{
It's a pain to write out manually the required "argument-list"
approach for the use of \code{under_construction}, but this
function can be used to fix that.
}

