% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uc_help_call.R
\name{uc_help_call}
\alias{uc_help_call}
\title{Help function for \code{\link{under_construction}}, using a call}
\usage{
uc_help_call(.uc_call, .save_file = NULL, high_details_level = FALSE,
  clean_workspace = FALSE, script_files = "")
}
\arguments{
\item{.uc_call}{The call we want to investigate, the name of the
function will be extracted from this.}

\item{.save_file}{A character string, to be used for the name of
the file to which \code{.uc_call} will be saved.  The default
value \code{NULL} will trigger the creation of a file name
based on the function-part of \code{.uc_call}.}

\item{high_details_level}{Logic value, default \code{FALSE}, to be
delivered to \code{under_construction}.}

\item{clean_workspace}{Logic value, default \code{FALSE}, to be
delivered to \code{under_construction}.}

\item{script_files}{Character vector, default \code{""} (i.e. an
empty string), to be delivered to \code{under_construction}.}
}
\value{
This function saves \code{.uc_call} to file and gives a
    line of code needed to source it back.  In addition it returns
    three chunks of code to simplify the setup when working with
    the function \code{under_construction}, i.e. it will create an
    argument list containing all the arguments of \code{.fun} with
    values from \code{.uc_call} and default values included, and
    this list will then be used to feed arguments to the
    \code{dotsMethods} argument of \code{under_construction}.  Some
    code needed for an inspection of the result is also included.
}
\description{
A help function similar to \code{\link{uc_help}}, but based on a
call (typically captured inside of some other function under
investigation).
}

